resolved #57
