+++ lib/app/shared/config/config.dart
@@ -1,0 +1,9 @@
+import 'package:corona_data/app/shared/utils/theme/constants.dart';
+import 'package:flutter/material.dart';
+
+final Map<String, ThemeData> themes = {
+  "light": themeLight,
+  "dark": themeDark,
+  "dracula": themeLight,
+};
+
