+++ lib/app/shared/utils/constants.dart
@@ -3,13 +3,7 @@ import 'package:google_maps_flutter/google_maps_flutter.dart';
 
 const String BASE_URL = "https://corona.lmao.ninja/v2";
 
-class ThemeColors{
-  const ThemeColors._();
-  static const Color success = Colors.green;
-  static const Color error = Colors.red;
-  static const Color warning = Colors.orange;
-  static const Color info = Colors.blue;
-}
+
 
 class AnimationSizes{
   static const double small=50;
